backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Supabase
supabase==2.0.0

# Hugging Face Embeddings
aiohttp==3.9.1
requests==2.31.0
tenacity==8.2.3

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.24.3

# Rate Limiting & Security
slowapi==0.1.9
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0